UI + Inventory + Interaction Systems + More
Added a Basic UI system with visible centre dot and Hotbar for items. Created an inventory system with a Item class with item IDs, Names, Descriptions, Stack Amounts And Tags, Inventory slot class with required methods and current amount of items in slot and create a inventory system class to integrate the inventory slots with required methods. Create an Interaction System with camera raycasting and a virtual interactable class as a basis for iteractive objects and create interactive item and interactive door script. Also normalised the movement vector before applying speed to keep WD, WA, SD, SA movement is the same speed. Create temporary materials for colours and updated the testing scene to test implementations.